+++ src/components/dataList.js
@@ -246,10 +246,21 @@
           }
         }, [loading]);
 
+        const handleClick = (event, context) => {
+          B.triggerEvent('OnItemClick', event, context);
+        };
+
+        const Wrapper = type === 'inline' ? 'span' : 'div';
         const Looper = results => {
           const rows = results.map(item => (
             <ModelProvider key={item.id} value={item} id={model}>
-              <B.InteractionScope>{children}</B.InteractionScope>
+              <B.InteractionScope model={model}>
+                {context => (
+                  <Wrapper onClick={event => handleClick(event, context)}>
+                    {children}
+                  </Wrapper>
+                )}
+              </B.InteractionScope>
             </ModelProvider>
           ));