+++ CHANGELOG.md
@@ -1,3 +1,48 @@
+## [1.72.1](https://github.com/bettyblocks/material-ui-component-set/compare/v1.72.0...v1.72.1) (2020-12-21)
+
+
+### Bug Fixes
+
+* object variable name based on model ([356fc96](https://github.com/bettyblocks/material-ui-component-set/commit/356fc96e7ed67a75b26b25b416b247fbb64685a2))
+
+# [1.72.0](https://github.com/bettyblocks/material-ui-component-set/compare/v1.71.1...v1.72.0) (2020-12-21)
+
+
+### Bug Fixes
+
+* select prefab for list prop ([a27588c](https://github.com/bettyblocks/material-ui-component-set/commit/a27588ca523af597054ff4ddb3a8b0e9d3367836))
+
+
+### Features
+
+* create form prefab ([cbb728d](https://github.com/bettyblocks/material-ui-component-set/commit/cbb728d97eabf2870d1ee5fcad73c74b6c4761b2))
+
+## [1.71.1](https://github.com/bettyblocks/material-ui-component-set/compare/v1.71.0...v1.71.1) (2020-12-17)
+
+
+### Bug Fixes
+
+* handle rowclick in datatable with outside variables ([adfc75c](https://github.com/bettyblocks/material-ui-component-set/commit/adfc75c4f6a194d688b31041059f663223cd0dbc))
+
+# [1.71.0](https://github.com/bettyblocks/material-ui-component-set/compare/v1.70.0...v1.71.0) (2020-12-17)
+
+
+### Features
+
+* add row click context to data list and data table ([322b691](https://github.com/bettyblocks/material-ui-component-set/commit/322b6919978e01482ff38e6773121500baafbc81))
+
+# [1.70.0](https://github.com/bettyblocks/material-ui-component-set/compare/v1.69.2...v1.70.0) (2020-12-16)
+
+
+### Bug Fixes
+
+* update cli ([0a75419](https://github.com/bettyblocks/material-ui-component-set/commit/0a754199dffea01c121ab22745f9ebf5e9c63a3d))
+
+
+### Features
+
+* update form prefab ([b18f82a](https://github.com/bettyblocks/material-ui-component-set/commit/b18f82aff6da4a22e8c46c42f5eaeb110c295db6))
+
 ## [1.69.2](https://github.com/bettyblocks/material-ui-component-set/compare/v1.69.1...v1.69.2) (2020-12-15)